+++ src/Controllers/SquareMeterController.ts
@@ -1,13 +1,22 @@
 import * as express from 'express'
+import { ISquerMeterData } from 'Models/dataTypings'
+import SquareMeterModel from 'Models/SquareMeterModel'
 import { BaseController } from './BaseController'
 
-export class SquareMetercontroller extends BaseController {
+export class SquareMeterController extends BaseController {
+	private db: typeof SquareMeterModel
+
+	constructor(_dataBase: typeof SquareMeterModel) {
+		super()
+		this.db = _dataBase
+	}
+
 	protected async execute(
 		req: express.Request,
 		res: express.Response,
-	): Promise<any> {
+	): Promise<void> {
 		try {
-			return this.ok<any>(res, { value: 10000 })
+			this.ok<ISquerMeterData>(res, await this.db.get())
 		} catch (error) {
 			this.fail(res, error.toString())
 		}