+++ src/Routes/index.ts
@@ -1,10 +1,11 @@
-import { SquareMetercontroller } from '../Controllers/SquareMeterController'
+import { SquareMeterController } from '../Controllers/SquareMeterController'
 import { Router } from 'express'
+import SquareMeterModel from '../Models/SquareMeterModel'
 
 const squareMeterRouter: Router = Router()
 
 squareMeterRouter.get('/', (req, res) =>
-	new SquareMetercontroller().get(req, res),
+	new SquareMeterController(SquareMeterModel).get(req, res),
 )
 
 export { squareMeterRouter }