+++ src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const DATABASE_FILE = 'src/DataBase/db.json'